+++ src/Data/Morpheus/Schema/TypeKind.hs
@@ -6,13 +6,12 @@ module Data.Morpheus.Schema.TypeKind
   ( TypeKind(..)
   ) where
 
-import           Data.Morpheus.Kind          (ENUM, KIND)
-import           Data.Morpheus.Types.GQLType (GQLType (__typeName, __typeVisibility))
+import           Data.Morpheus.Kind          (ENUM)
+import           Data.Morpheus.Types.GQLType (GQLType (KIND, __typeName, __typeVisibility))
 import           GHC.Generics
 
-type instance KIND TypeKind = ENUM
-
 instance GQLType TypeKind where
+  type KIND TypeKind = ENUM
   __typeName = const "__TypeKind"
   __typeVisibility = const False