+++ test/Feature/Schema/A2.hs
@@ -1,17 +1,17 @@
-{-# LANGUAGE DeriveAnyClass #-}
-{-# LANGUAGE DeriveGeneric  #-}
-{-# LANGUAGE TypeFamilies   #-}
+{-# LANGUAGE DeriveGeneric #-}
+{-# LANGUAGE TypeFamilies  #-}
 
 module Feature.Schema.A2
   ( A(..)
   ) where
 
-import           Data.Morpheus.Kind  (KIND, OBJECT)
+import           Data.Morpheus.Kind  (OBJECT)
 import           Data.Morpheus.Types (GQLType (..))
 import           GHC.Generics        (Generic)
 
-type instance KIND A = OBJECT
-
 newtype A = A
   { bla :: Int
-  } deriving (Generic, GQLType)
+  } deriving (Generic)
+
+instance GQLType A where
+  type KIND A = OBJECT